+++ hippynn/graphs/gops.py
@@ -8,6 +8,7 @@ from .nodes.base import InputNode, MultiNode
 from .nodes.base.algebra import ValueNode
 from .nodes.base.node_functions import NodeNotFound, NodeOperationError
 from .indextypes import soft_index_type_coercion
+
 from . import get_connected_nodes, find_unique_relative
 from ..tools import is_equal_state_dict
 
@@ -475,3 +476,45 @@ def vacuum_outputs(node_list, species_set):
         replace_node_with_constant(pairs.pair_coord, empty_displacement, "empty_displacements")
 
     return new_nodes
+
+def swap_pairfinders(node_or_nodes, new_pairfinder, new_node_name=None, cell_node=None, module_kwargs={}):
+    """
+    Finds and replaces existing PairIndexer node with a new one, potentially adjusting its parent nodes
+    if needed.
+
+    :param node_or_nodes: the PairIndexer node to be replaced, or a node or list of nodes connected
+    to the unique PairIndexer to be replaced
+    :param new_pairfinder: class of new PairIndexer 
+    :param new_node_name: name for new PairIndexer node, if None the name of the replaced PairIndexer
+    will be used, defaults to None
+    :param cell_node: should be specified if new PairIndexer requires a CellNode and one does not currently
+    exist in computational graph, if None a search of existing nodes will be conducted if a CellNode is 
+    needed, defaults to None
+    :param module_kwargs: arguments to feed into the new PairIndexer constructor, defaults to {}
+
+    .. Note::
+    
+        If this function is used to add a CellNode to the computational graph, any existing GraphModule or 
+        Predictor will need to be reinitialized.
+    """
+
+    from .nodes.tags import PairIndexer, Positions, Species
+    from .nodes.inputs import CellNode
+
+    if isinstance(node_or_nodes, PairIndexer):
+        old_pf = node_or_nodes
+    else:
+        old_pf = find_unique_relative(node_or_nodes, PairIndexer)
+
+    positions = find_unique_relative(old_pf, Positions)
+    species = find_unique_relative(old_pf, Species)
+
+    new_node_name = (new_node_name or old_pf.name)
+
+    try:
+        new_pf = new_pairfinder(new_node_name, parents=(positions, species), **module_kwargs)
+    except RuntimeError:
+        cell = (cell_node or find_unique_relative(old_pf, CellNode))
+        new_pf = new_pairfinder(new_node_name, parents=(positions, species, cell), **module_kwargs)
+    
+    replace_node(old_pf, new_pf, disconnect_old=True)